test: separate existing sns tests into general tests and s3 event tests (#803)
