+++ __tests__/api-sns-with-s3-event.test.ts
@@ -1,8 +1,3 @@
-// Mock HTTP wrapper for outbound requests BEFORE any imports
-jest.mock("../app/utils/http", () => {
-  return { httpGet: jest.fn() };
-});
-
 // Set up AWS resource configuration BEFORE any other imports
 const TEST_S3_BUCKET = "hca-atlas-tracker-data-dev";
 const TEST_GUT_ATLAS_ID = "550e8400-e29b-41d4-a716-446655440000";
@@ -31,20 +26,11 @@ import {
   countSourceDatasets,
   resetDatabase,
 } from "../testing/db-utils";
-import {
-  withConsoleErrorHiding,
-  withConsoleMessageHiding,
-} from "../testing/utils";
+import { withConsoleMessageHiding } from "../testing/utils";
 
 // Additional imports for direct service testing
 import { updateSourceDataset } from "../app/services/s3-notification";
 import { InvalidOperationError } from "../app/utils/api-handler";
-
-// Retrieve the mock function created in the jest.mock factory above
-const { httpGet } = jest.requireMock("../app/utils/http") as {
-  httpGet: jest.Mock;
-};
-const mockHttpGet = httpGet as jest.Mock;
 
 // Test file path constants
 const TEST_PATH_SEGMENTS = {
@@ -253,19 +239,7 @@ afterAll(() => {
   endPgPool();
 });
 
-describe(TEST_ROUTE, () => {
-  // HTTP Method Validation Tests
-  it("returns error 405 for non-POST request", async () => {
-    const { req, res } = httpMocks.createMocks<NextApiRequest, NextApiResponse>(
-      {
-        method: METHOD.GET,
-      }
-    );
-
-    await snsHandler(req, res);
-
-    expect(res.statusCode).toBe(405);
-  });
+describe(`${TEST_ROUTE} (S3 event)`, () => {
   it("rejects S3 record when eventTime is missing", async () => {
     // Start with a valid event and then remove eventTime to simulate missing timestamp
     const s3Event = createS3Event({
@@ -306,24 +280,6 @@ describe(TEST_ROUTE, () => {
       TEST_FILE_PATHS.SOURCE_DATASET_TEST,
     ]);
     expect(fileRows.rows).toHaveLength(0);
-  });
-
-  // SNS Message Structure Validation Tests
-  it("returns error 400 for invalid SNS message payload", async () => {
-    const { req, res } = httpMocks.createMocks<NextApiRequest, NextApiResponse>(
-      {
-        body: {
-          invalid: "payload",
-        },
-        method: METHOD.POST,
-      }
-    );
-
-    await withConsoleErrorHiding(async () => {
-      await snsHandler(req, res);
-    });
-
-    expect(res.statusCode).toBe(400);
   });
 
   // SHA256 Integrity Validation Tests
@@ -737,129 +693,6 @@ describe(TEST_ROUTE, () => {
     );
   });
 
-  // Security and Authentication Tests
-  it("rejects SNS messages with invalid signatures", async () => {
-    const s3Event = createS3Event({
-      etag: "invalid-signature-test",
-      eventTime: TEST_TIMESTAMP_ALT,
-      key: TEST_FILE_PATHS.SOURCE_DATASET_AUTH,
-      size: 256000,
-      versionId: "auth-test-version",
-    });
-
-    const snsMessageWithInvalidSignature = createSNSMessage({
-      messageId: "auth-test-message",
-      s3Event,
-      signature: TEST_SIGNATURE_INVALID,
-      timestamp: TEST_TIMESTAMP_ALT,
-    });
-
-    const { req, res } = httpMocks.createMocks<NextApiRequest, NextApiResponse>(
-      {
-        body: snsMessageWithInvalidSignature,
-        method: "POST",
-      }
-    );
-
-    await withConsoleErrorHiding(async () => {
-      await snsHandler(req, res);
-    });
-
-    // Should reject with 401 Unauthorized due to invalid signature
-    expect(res.statusCode).toBe(401);
-    expect(JSON.parse(res._getData())).toEqual({
-      message: "SNS signature validation failed",
-    });
-
-    // Verify no file was saved to database
-    const fileRows = await query(SQL_QUERIES.SELECT_FILE_BY_BUCKET_AND_KEY, [
-      TEST_S3_BUCKET,
-      TEST_FILE_PATHS.SOURCE_DATASET_AUTH,
-    ]);
-
-    expect(fileRows.rows).toHaveLength(0);
-  });
-
-  // Test data for SNS validator edge cases
-  const snsValidatorEdgeCases = [
-    {
-      expectedMessage: "SNS validation returned no message",
-      mockBehavior: (
-        message: Record<string, unknown>,
-        callback: (err: Error | null, result?: Record<string, unknown>) => void
-      ): void => callback(null, undefined),
-      size: 128000,
-      testId: "no-validated-message",
-      testName:
-        "rejects SNS messages when validator returns no validated message",
-    },
-  ];
-
-  test.each(snsValidatorEdgeCases)(
-    "$testName",
-    async ({ expectedMessage, mockBehavior, size, testId }) => {
-      // Temporarily override the mock with specific behavior
-      const MockedMessageValidator = jest.mocked(
-        jest.requireMock(TEST_MODULE_SNS_VALIDATOR)
-      );
-      MockedMessageValidator.mockImplementation(() => ({
-        validate: jest.fn(mockBehavior),
-      }));
-
-      const s3Event = createS3Event({
-        etag: `${testId}-test`,
-        eventTime: TEST_TIMESTAMP_ALT,
-        key: TEST_FILE_PATHS.SOURCE_DATASET_AUTH,
-        size,
-        versionId: `${testId}-version`,
-      });
-
-      const snsMessage = createSNSMessage({
-        messageId: `${testId}-test`,
-        s3Event,
-        signature: TEST_SIGNATURE_VALID,
-        timestamp: TEST_TIMESTAMP_ALT,
-      });
-
-      const { req, res } = httpMocks.createMocks<
-        NextApiRequest,
-        NextApiResponse
-      >({
-        body: snsMessage,
-        method: "POST",
-      });
-
-      await withConsoleErrorHiding(async () => {
-        await snsHandler(req, res);
-      });
-
-      // Should reject with 400 Bad Request
-      expect(res.statusCode).toBe(400);
-      expect(JSON.parse(res._getData())).toEqual({
-        message: expectedMessage,
-      });
-
-      // Verify no file was saved to database
-      const fileRows = await query(SQL_QUERIES.SELECT_FILE_BY_BUCKET_AND_KEY, [
-        TEST_S3_BUCKET,
-        TEST_FILE_PATHS.SOURCE_DATASET_AUTH,
-      ]);
-
-      expect(fileRows.rows).toHaveLength(0);
-
-      // Restore original mock
-      MockedMessageValidator.mockImplementation(() => ({
-        validate: jest.fn((message, callback) => {
-          if (message.Signature === TEST_SIGNATURE_INVALID) {
-            callback(new Error(TEST_ERROR_MESSAGE_INVALID_SIGNATURE));
-            return;
-          }
-          callback(null, message);
-        }),
-      }));
-    }
-  );
-
   it("rejects notifications with ETag mismatches for existing files", async () => {
     const s3Event = createS3Event({
       etag: "original-etag-12345",
@@ -1092,42 +925,6 @@ describe(TEST_ROUTE, () => {
     expect(latestOnly.rows[0].etag).toBe(
       "ooo-version2-etag-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
     );
-  });
-
-  it("rejects notifications from unauthorized SNS topics", async () => {
-    await withConsoleMessageHiding(async () => {
-      const s3Event = createS3Event({
-        etag: "d41d8cd98f00b204e9800998ecf8427e",
-        key: "test-file.h5ad",
-        size: 1024000,
-        versionId: TEST_VERSION_IDS.DEFAULT,
-      });
-
-      const unauthorizedSnsMessage = createSNSMessage({
-        messageId: "unauthorized-topic-test",
-        s3Event,
-        signature: TEST_SIGNATURE,
-        subject: SNS_MESSAGE_DEFAULTS.SUBJECT,
-        timestamp: TEST_TIMESTAMP,
-        topicArn: UNAUTHORIZED_TOPIC_ARN,
-      });
-
-      const { req, res } = httpMocks.createMocks<
-        NextApiRequest,
-        NextApiResponse
-      >({
-        body: unauthorizedSnsMessage,
-        method: METHOD.POST,
-      });
-
-      const handler = (await import("../pages/api/sns")).default;
-      await handler(req, res);
-
-      expect(res._getStatusCode()).toBe(403);
-      expect(JSON.parse(res._getData())).toEqual({
-        message: `Unauthorized SNS topic: ${UNAUTHORIZED_TOPIC_ARN}`,
-      });
-    });
   });
 
   it("rejects notifications from unauthorized S3 buckets", async () => {
@@ -1631,221 +1428,6 @@ describe(TEST_ROUTE, () => {
     expect(res.statusCode).toBe(400);
     expect(JSON.parse(res._getData())).toEqual({
       message: "Expected exactly 1 S3 record, but received 2 records",
-    });
-  });
-
-  // SNS Test Constants (shared across subscription tests)
-  const SUBSCRIPTION_CONFIRMATION_SUBJECT =
-    "AWS Notification - Subscription Confirmation";
-  const SUBSCRIPTION_CONFIRMATION_MESSAGE =
-    "You have chosen to subscribe to the topic";
-  const SNS_SIGNATURE_VERSION = "1";
-  const SIGNING_CERT_URL =
-    "https://sns.us-east-1.amazonaws.com/SimpleNotificationService-test.pem";
-  const SUBSCRIPTION_CONFIRMATION_TYPE = "SubscriptionConfirmation";
-  const SUBSCRIBE_URL_BASE =
-    "https://sns.us-east-1.amazonaws.com/?Action=ConfirmSubscription&TopicArn=";
-  const SUBSCRIBE_URL = `${SUBSCRIBE_URL_BASE}test&Token=test-token`;
-  const UNAUTHORIZED_TOPIC_ARN =
-    "arn:aws:sns:us-east-1:123456789012:unauthorized-topic";
-  const UNSUBSCRIBE_MESSAGE = "You have unsubscribed from the topic";
-
-  describe("SNS SubscriptionConfirmation handling", () => {
-    beforeEach(() => {
-      // Reset HTTP wrapper mock before each test
-      mockHttpGet.mockClear();
-      mockHttpGet.mockResolvedValue({ ok: true });
-    });
-
-    describe("SubscriptionConfirmation", () => {
-      it("handles valid subscription confirmation with proper topic authorization", async () => {
-        const subscriptionConfirmationMessage = {
-          Message: SUBSCRIPTION_CONFIRMATION_MESSAGE,
-          MessageId: "subscription-test-message-id",
-          Signature: TEST_SIGNATURE,
-          SignatureVersion: SNS_SIGNATURE_VERSION,
-          SigningCertURL: SIGNING_CERT_URL,
-          Subject: SUBSCRIPTION_CONFIRMATION_SUBJECT,
-          SubscribeURL: SUBSCRIBE_URL,
-          Timestamp: TEST_TIMESTAMP,
-          TopicArn: TEST_AWS_CONFIG.sns_topics[0], // Use authorized topic
-          Type: SUBSCRIPTION_CONFIRMATION_TYPE,
-        };
-
-        const { req, res } = httpMocks.createMocks<
-          NextApiRequest,
-          NextApiResponse
-        >({
-          body: subscriptionConfirmationMessage,
-          method: METHOD.POST,
-        });
-
-        const handler = (await import("../pages/api/sns")).default;
-        await withConsoleMessageHiding(async () => {
-          await handler(req, res);
-        });
-
-        expect(res.statusCode).toBe(200);
-        // HTTP call verification: The 200 response indicates the SubscribeURL was successfully called
-      });
-
-      it("rejects subscription confirmation from unauthorized topic", async () => {
-        const unauthorizedSubscriptionMessage = {
-          Message: SUBSCRIPTION_CONFIRMATION_MESSAGE,
-          MessageId: "unauthorized-subscription-message-id",
-          Signature: TEST_SIGNATURE,
-          SignatureVersion: SNS_SIGNATURE_VERSION,
-          SigningCertURL: SIGNING_CERT_URL,
-          Subject: SUBSCRIPTION_CONFIRMATION_SUBJECT,
-          SubscribeURL: SUBSCRIBE_URL,
-          Timestamp: TEST_TIMESTAMP,
-          TopicArn: UNAUTHORIZED_TOPIC_ARN,
-          Type: SUBSCRIPTION_CONFIRMATION_TYPE,
-        };
-
-        const { req, res } = httpMocks.createMocks<
-          NextApiRequest,
-          NextApiResponse
-        >({
-          body: unauthorizedSubscriptionMessage,
-          method: METHOD.POST,
-        });
-
-        const handler = (await import("../pages/api/sns")).default;
-        await withConsoleMessageHiding(async () => {
-          await handler(req, res);
-        });
-
-        expect(res.statusCode).toBe(403);
-        expect(JSON.parse(res._getData())).toEqual({
-          message: `Unauthorized SNS topic: ${UNAUTHORIZED_TOPIC_ARN}`,
-        });
-      });
-
-      it("rejects subscription confirmation with missing SubscribeURL", async () => {
-        const subscriptionMessageWithoutURL = {
-          Message: SUBSCRIPTION_CONFIRMATION_MESSAGE,
-          MessageId: "missing-url-test-message-id",
-          Signature: TEST_SIGNATURE,
-          SignatureVersion: SNS_SIGNATURE_VERSION,
-          SigningCertURL: SIGNING_CERT_URL,
-          Subject: SUBSCRIPTION_CONFIRMATION_SUBJECT,
-          // SubscribeURL is intentionally missing
-          Timestamp: TEST_TIMESTAMP,
-          TopicArn: TEST_AWS_CONFIG.sns_topics[0],
-          Type: SUBSCRIPTION_CONFIRMATION_TYPE,
-        };
-
-        const { req, res } = httpMocks.createMocks<
-          NextApiRequest,
-          NextApiResponse
-        >({
-          body: subscriptionMessageWithoutURL,
-          method: METHOD.POST,
-        });
-
-        const handler = (await import("../pages/api/sns")).default;
-        await withConsoleMessageHiding(async () => {
-          await handler(req, res);
-        });
-
-        expect(res.statusCode).toBe(400);
-        expect(JSON.parse(res._getData())).toEqual({
-          message: "SubscribeURL is required for subscription confirmation",
-        });
-      });
-    });
-
-    describe("UnsubscribeConfirmation", () => {
-      it("handles valid unsubscribe confirmation with proper topic authorization", async () => {
-        const unsubscribeConfirmationMessage = {
-          Message: UNSUBSCRIBE_MESSAGE,
-          MessageId: "unsubscribe-test-message-id",
-          Signature: TEST_SIGNATURE,
-          SignatureVersion: SNS_SIGNATURE_VERSION,
-          SigningCertURL: SIGNING_CERT_URL,
-          Subject: "AWS Notification - Unsubscribe Confirmation",
-          Timestamp: TEST_TIMESTAMP,
-          TopicArn: TEST_AWS_CONFIG.sns_topics[0], // Use authorized topic
-          Type: "UnsubscribeConfirmation",
-        };
-
-        const { req, res } = httpMocks.createMocks<
-          NextApiRequest,
-          NextApiResponse
-        >({
-          body: unsubscribeConfirmationMessage,
-          method: METHOD.POST,
-        });
-
-        const handler = (await import("../pages/api/sns")).default;
-        await withConsoleMessageHiding(async () => {
-          await handler(req, res);
-        });
-
-        expect(res.statusCode).toBe(200);
-      });
-
-      it("rejects unsubscribe confirmation from unauthorized topic", async () => {
-        const unauthorizedUnsubscribeMessage = {
-          Message: UNSUBSCRIBE_MESSAGE,
-          MessageId: "unauthorized-unsubscribe-message-id",
-          Signature: TEST_SIGNATURE,
-          SignatureVersion: SNS_SIGNATURE_VERSION,
-          SigningCertURL: SIGNING_CERT_URL,
-          Subject: "AWS Notification - Unsubscribe Confirmation",
-          Timestamp: TEST_TIMESTAMP,
-          TopicArn: UNAUTHORIZED_TOPIC_ARN,
-          Type: "UnsubscribeConfirmation",
-        };
-
-        const { req, res } = httpMocks.createMocks<
-          NextApiRequest,
-          NextApiResponse
-        >({
-          body: unauthorizedUnsubscribeMessage,
-          method: METHOD.POST,
-        });
-
-        const handler = (await import("../pages/api/sns")).default;
-        await withConsoleMessageHiding(async () => {
-          await handler(req, res);
-        });
-
-        expect(res.statusCode).toBe(403);
-        expect(JSON.parse(res._getData())).toEqual({
-          message: `Unauthorized SNS topic: ${UNAUTHORIZED_TOPIC_ARN}`,
-        });
-      });
-
-      it("handles unsubscribe confirmation without Subject field", async () => {
-        const unsubscribeMessageWithoutSubject = {
-          Message: UNSUBSCRIBE_MESSAGE,
-          MessageId: "no-subject-unsubscribe-message-id",
-          Signature: TEST_SIGNATURE,
-          SignatureVersion: SNS_SIGNATURE_VERSION,
-          SigningCertURL: SIGNING_CERT_URL,
-          // Subject is intentionally missing
-          Timestamp: TEST_TIMESTAMP,
-          TopicArn: TEST_AWS_CONFIG.sns_topics[0],
-          Type: "UnsubscribeConfirmation",
-        };
-
-        const { req, res } = httpMocks.createMocks<
-          NextApiRequest,
-          NextApiResponse
-        >({
-          body: unsubscribeMessageWithoutSubject,
-          method: METHOD.POST,
-        });
-
-        const handler = (await import("../pages/api/sns")).default;
-        await withConsoleMessageHiding(async () => {
-          await handler(req, res);
-        });
-
-        expect(res.statusCode).toBe(200);
-      });
     });
   });
 });
